Move platform check to CLI entry point
At the moment the platform check is done when the policy module gets
imported. This makes it difficult to run the test suite on non-Linux
platforms. Moving the check to the CLI entry point allows the tests to
run and retains the current behavior and exits on non-Linux platforms
when running the tool.

Note that in a previous commit I accidentally commented out the check,
that's why it's shown commented in this diff.