Rename test_policy integration test
As the new `test_policy.py` unit test file name clashes with the
original integration test's name this renames the integration test.
`pytest` does not allow the same file name across different paths in the
test suite.

As the integration test is more about validating the policies stored in
the project, I feel it makes more sense to rename it to
`test_policy_files`.