Vendor wheel 0.36.2 (#275)
* Vendor wheel 0.36.2

`wheel` does not provide a public API yet. In order not to be impacted by any internal refactoring/changes in the `wheel` package, the version of the `wheel` package is pinned by `auditwheel`. [This causes some nuisance for users of `auditwheel`](https://github.com/pypa/auditwheel/issues/274).
This commit vendors `wheel` to avoid such nuisance and while waiting for a public API in `wheel`.

* Omit coverage report for vendored packages

* Exclude vendored packages from flake8 check

* Fix flake8 issue

* Exclude vendored packages from mypy check