Only create a single repaired wheel
When repairing wheels, the result can be compliant with a lower-level
manylinux spec and in this case auditwheel creates another .whl file for
the lower spec as well. This is unnecessary as the lower spec can always
be used on newer compliant platforms. By only producing a single wheel,
we can probably save some space on PyPI as automated build scripts won't
upload multiple copies of the same wheel.

This also avoids an extra call to repair_wheel which can speed things up
for larger libraries.