Rework auditwheel show checks in test_manylinux.py
With multiple manylinux images, it's not always easy to guess what the repaired wheel will be compatible with.
Add a helper for the auditwheel show check that allows to relax the constraint on the expected policy as long as it's compatible with the expected one.
The `strict` argument of the helper allows to choose wether the check shall be relaxed or not.