Remove globals in policy module
The policy module creates several globals on the module level, which
makes it very difficult to test code which interacts with the policies.

This patch creates a `WheelPolicies` class which encapsulates the
policy processing code and exposes functions to query the parsed
policies. Unfortunately, a lot of functions directly operate on
policies and these all need a new argument to take the new policy
object. This change should make testing substantially easier as policies
can be easily created on the fly and passed down to functions which
require these.