feat: add GNU_PROPERTY_X86_ISA_1_NEEDED detection (#535)
* chore: use a dataclass for ldd result

* feat: add GNU_PROPERTY_X86_ISA_1_NEEDED detection

ISA extensions usage is not defined by a PEP yet.

This first implementation fails to repair the wheel if the usage of x86-64-v[2-4] is required.

The check can be disabled with `--disable-isa-ext-check`.

The detection being related to a declaration when building, it will not detect the requirement for binaries where the declaration is missing.

All executables built on a manylinux_2_34  image will be detected as x86-64-v2.