Add libatomic.so.1 to lib_whitelist
On RISC-V, GCC 12 and earlier require libatomic support for 1-byte and 2-byte operations, which have been a long and painful story. [1][2][3][4]

Users' machine definitely will have the libatomic library to make softwares work, so adding this should be safe.

This seems also to be an regression on cpython 3.13 with more architectures than riscv64. [5]

[1] https://github.com/riscvarchive/riscv-gcc/issues/12
[2] https://github.com/riscvarchive/riscv-gcc/issues/337
[3] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=86005
[4] https://github.com/python/cpython/issues/112779
[5] https://github.com/python/cpython/issues/127708